backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
google-generativeai==0.3.2
requests==2.31.0